+++ src/components/heatmap/Heatmap.test.js
@@ -16,12 +16,14 @@ describe('<Heatmap/>', () => {
         theme="dark"
         width={100}
         height={100}
+        colormap="plasma"
+        colormapRange={[0.0, 1.0]}
         expressionMatrix={expressionMatrix}
         cellColors={cellColors}
         transpose
         viewState={{ zoom: 0, target: [0, 0] }}
       />,
     );
-    expect(wrapper.find('#deckgl-wrapper').length).toEqual(1);
+    expect(wrapper.find('#deckgl-overlay-heatmap-0-wrapper').length).toEqual(1);
   });
 });