+++ src/components/hooks.js
@@ -1,5 +1,5 @@
 import {
-  useRef, useState, useEffect, useCallback,
+  useRef, useState, useEffect, useCallback, useMemo,
 } from 'react';
 import debounce from 'lodash/debounce';
 import { useGridResize, useEmitGridResize } from '../app/state/hooks';
@@ -125,9 +125,10 @@ export function useDeckCanvasSize() {
  * (outside a function component / render function),
  * otherwise strange bugs may occur.
  * @returns {array} An array
- * [isReady, setItemIsReady, resetReadyItems]
+ * [isReady, setItemIsReady, setItemIsNotReady, resetReadyItems]
  * where isReady is the boolean value,
  * setItemIsReady marks one item as ready,
+ * setItemIsNotReady marks one item as not ready,
  * and resetReadyItem marks all items as waiting.
  */
 export function useReady(supportedItems) {
@@ -143,6 +144,15 @@ export function useReady(supportedItems) {
     });
   }, [setWaiting]);
 
+  const setItemIsNotReady = useCallback((notReadyItem) => {
+    setWaiting((waitingItems) => {
+      const nextWaitingItems = [...waitingItems, notReadyItem];
+      // eslint-disable-next-line no-console
+      console.log(`waiting on ${nextWaitingItems.length}: ${JSON.stringify(nextWaitingItems)}`);
+      return nextWaitingItems;
+    });
+  }, [setWaiting]);
+
   const resetReadyItems = useCallback(() => {
     setWaiting(items);
     // eslint-disable-next-line no-console
@@ -151,7 +161,7 @@ export function useReady(supportedItems) {
 
   const isReady = waiting.length === 0;
 
-  return [isReady, setItemIsReady, resetReadyItems];
+  return [isReady, setItemIsReady, setItemIsNotReady, resetReadyItems];
 }
 
 /**
@@ -209,4 +219,31 @@ export function useClosestVitessceContainerSize(ref) {
   }, [ref]);
 
   return [width, height];
+}
+
+export function useExpressionValueGetter({ attrs, expressionData }) {
+  // Get a mapping from cell ID to row index in the gene expression matrix.
+  const cellIdMap = useMemo(() => {
+    const result = {};
+    if (attrs && attrs.rows) {
+      // eslint-disable-next-line no-plusplus
+      for (let i = 0; i < attrs.rows.length; i++) {
+        result[attrs.rows[i]] = i;
+      }
+    }
+    return result;
+  }, [attrs]);
+
+  // Set up a getter function for gene expression values, to be used
+  // by the DeckGL layer to obtain values for instanced attributes.
+  const getExpressionValue = useCallback((entry) => {
+    const cellId = entry[0];
+    if (cellIdMap && expressionData && expressionData[0]) {
+      const cellIndex = cellIdMap[cellId];
+      const val = expressionData[0][cellIndex];
+      return val;
+    }
+    return 0;
+  }, [cellIdMap, expressionData]);
+  return getExpressionValue;
 }