+++ src/components/layer-controller/LayerControllerSubscriber.js
@@ -51,6 +51,7 @@ const LayerControllerMemoized = React.memo(
       handleRasterLayerRemove,
       disable3d,
       globalDisable3d,
+      disableChannelsIfRgbDetected,
       layerIs3DIndex,
       setZoom,
       setTargetX,
@@ -149,6 +150,7 @@ const LayerControllerMemoized = React.memo(
                       && layerIs3DIndex !== -1
                       && layerIs3DIndex !== i
                     }
+                    disableChannelsIfRgbDetected={disableChannelsIfRgbDetected}
                     rasterLayersCallbacks={rasterLayersCallbacks}
                     setRasterLayerCallback={setRasterLayerCallback}
                     setViewState={({
@@ -204,6 +206,7 @@ function LayerControllerSubscriber(props) {
     title = 'Spatial Layers',
     disable3d,
     globalDisable3d,
+    disableChannelsIfRgbDetected,
   } = props;
 
   const loaders = useLoaders();
@@ -253,7 +256,12 @@ function LayerControllerSubscriber(props) {
   const [componentWidth, componentHeight] = useClosestVitessceContainerSize(layerControllerRef);
   const { height: windowHeight, width: windowWidth } = useWindowDimensions();
 
-  const [isReady, setItemIsReady, resetReadyItems] = useReady(
+  const [
+    isReady,
+    setItemIsReady,
+    setItemIsNotReady, // eslint-disable-line no-unused-vars
+    resetReadyItems,
+  ] = useReady(
     LAYER_CONTROLLER_DATA_TYPES,
   );
 
@@ -344,6 +352,7 @@ function LayerControllerSubscriber(props) {
       disable3d={disable3d}
       globalDisable3d={globalDisable3d}
       layerIs3DIndex={layerIs3DIndex}
+      disableChannelsIfRgbDetected={disableChannelsIfRgbDetected}
       setZoom={setZoom}
       setTargetX={setTargetX}
       setTargetY={setTargetY}