+++ src/components/layer-controller/LayerOptions.js
@@ -8,47 +8,16 @@ import Select from '@material-ui/core/Select';
 import Checkbox from '@material-ui/core/Checkbox';
 import { getDefaultInitialViewState } from '@hms-dbmi/viv';
 
-import { getBoundingCube, getMultiSelectionStats } from './utils';
+import {
+  getBoundingCube, getMultiSelectionStats, formatBytes,
+  getStatsForResolution,
+  canLoadResolution,
+} from './utils';
 import { COLORMAP_OPTIONS } from '../utils';
 import { DEFAULT_RASTER_DOMAIN_TYPE } from '../spatial/constants';
 import { StyledSelectionSlider } from './styles';
 
 const DOMAIN_OPTIONS = ['Full', 'Min/Max'];
-
-function formatBytes(bytes, decimals = 2) {
-  if (bytes === 0) return '0 Bytes';
-
-  const k = 1024;
-  const dm = decimals < 0 ? 0 : decimals;
-  const sizes = ['Bytes', 'KB', 'MB', 'GB'];
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-  // eslint-disable-next-line no-restricted-properties
-  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`;
-}
-
-const getStatsForResolution = (loader, resolution) => {
-  const { shape, labels } = loader[resolution];
-  const height = shape[labels.indexOf('y')];
-  const width = shape[labels.indexOf('x')];
-  const depth = shape[labels.indexOf('z')];
-  const depthDownsampled = Math.floor(depth / (2 ** resolution));
-  // Check memory allocation limits for Float32Array (used in XR3DLayer for rendering)
-  const totalBytes = 4 * height * width * depthDownsampled;
-  return {
-    height, width, depthDownsampled, totalBytes,
-  };
-};
-
-const canLoadResolution = (loader, resolution) => {
-  const { totalBytes } = getStatsForResolution(loader, resolution);
-  const maxHeapSize = window.performance?.memory
-    && window.performance?.memory?.jsHeapSizeLimit / 2;
-  const maxSize = maxHeapSize || (2 ** 31) - 1;
-  return totalBytes < maxSize;
-};
-
 
 /**
  * Wrapper for the dropdown that selects a colormap (None, viridis, magma, etc.).
@@ -357,7 +326,7 @@ function LayerOption({ name, inputId, children }) {
  * @prop {array} channels Current channel object for inferring the current global selection.
  * @prop {array} dimensions Currently available dimensions (channel, z, t etc.).
  * @prop {string} domainType One of Max/Min or Full (soon presets as well).
- * @prop {boolean} isRgb Whether or not the image is rgb (so we don't need colormap controllers).
+ * @prop {boolean} disableChannelsIfRgbDetected Whether or not we need colormap controllers if RGB.
  */
 function LayerOptions({
   colormap,
@@ -373,7 +342,7 @@ function LayerOptions({
   transparentColor,
   channels,
   domainType,
-  isRgb,
+  disableChannelsIfRgbDetected,
   shouldShowTransparentColor,
   shouldShowDomain,
   shouldShowColormap,
@@ -392,10 +361,15 @@ function LayerOptions({
   const { labels, shape } = Array.isArray(loader.data) ? loader.data[0] : loader.data;
   const hasDimensionsAndChannels = labels.length > 0 && channels.length > 0;
   const hasZStack = shape[labels.indexOf('z')] > 1;
+  // Only show volume button if we can actually view resolutions.
+  const hasViewableResolutions = Boolean(Array.from({
+    length: loader.data.length,
+  }).filter((_, res) => canLoadResolution(loader.data, res)).length);
   return (
     <Grid container direction="column" style={{ width: '100%' }}>
       {hasZStack
         && !disable3d
+        && hasViewableResolutions
         && (
         <VolumeDropdown
           loader={loader}
@@ -429,7 +403,7 @@ function LayerOptions({
           </LayerOption>
           ),
         )}
-      {!isRgb ? (
+      {!disableChannelsIfRgbDetected ? (
         <>
           {shouldShowColormap && (
             <Grid item>