+++ src/components/scatterplot/ScatterplotOptions.js
@@ -1,18 +1,27 @@
-import React from 'react';
+import React, { useCallback } from 'react';
+import debounce from 'lodash/debounce';
 import Checkbox from '@material-ui/core/Checkbox';
 import Slider from '@material-ui/core/Slider';
 import TableCell from '@material-ui/core/TableCell';
 import TableRow from '@material-ui/core/TableRow';
+import Select from '@material-ui/core/Select';
 import { capitalize } from '../../utils';
 import { useStyles } from '../shared-plot-options/styles';
 import OptionsContainer from '../shared-plot-options/OptionsContainer';
 import CellColorEncodingOption from '../shared-plot-options/CellColorEncodingOption';
+import { GLSL_COLORMAPS } from '../../layers/constants';
 
 export default function ScatterplotOptions(props) {
   const {
     observationsLabel,
     cellRadius,
     setCellRadius,
+    cellRadiusMode,
+    setCellRadiusMode,
+    cellOpacity,
+    setCellOpacity,
+    cellOpacityMode,
+    setCellOpacityMode,
     cellSetLabelsVisible,
     setCellSetLabelsVisible,
     cellSetLabelSize,
@@ -21,14 +30,30 @@ export default function ScatterplotOptions(props) {
     setCellSetPolygonsVisible,
     cellColorEncoding,
     setCellColorEncoding,
+    geneExpressionColormap,
+    setGeneExpressionColormap,
+    geneExpressionColormapRange,
+    setGeneExpressionColormapRange,
   } = props;
 
   const observationsLabelNice = capitalize(observationsLabel);
 
   const classes = useStyles();
 
+  function handleCellRadiusModeChange(event) {
+    setCellRadiusMode(event.target.value);
+  }
+
+  function handleCellOpacityModeChange(event) {
+    setCellOpacityMode(event.target.value);
+  }
+
   function handleRadiusChange(event, value) {
     setCellRadius(value);
+  }
+
+  function handleOpacityChange(event, value) {
+    setCellOpacity(value);
   }
 
   function handleLabelVisibilityChange(event) {
@@ -43,8 +68,25 @@ export default function ScatterplotOptions(props) {
     setCellSetPolygonsVisible(event.target.checked);
   }
 
+  function handleGeneExpressionColormapChange(event) {
+    setGeneExpressionColormap(event.target.value);
+  }
+
+  function handleColormapRangeChange(event, value) {
+    setGeneExpressionColormapRange(value);
+  }
+  const handleColormapRangeChangeDebounced = useCallback(
+    debounce(handleColormapRangeChange, 5, { trailing: true }),
+    [handleColormapRangeChange],
+  );
+
   return (
     <OptionsContainer>
+      <CellColorEncodingOption
+        observationsLabel={observationsLabel}
+        cellColorEncoding={cellColorEncoding}
+        setCellColorEncoding={setCellColorEncoding}
+      />
       <TableRow>
         <TableCell className={classes.labelCell}>
           {observationsLabelNice} Set Labels Visible
@@ -92,27 +134,116 @@ export default function ScatterplotOptions(props) {
         </TableCell>
       </TableRow>
       <TableRow>
+        <TableCell className={classes.labelCell} htmlFor="cell-radius-mode-select">
+          {observationsLabelNice} Radius Mode
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Select
+            native
+            className={classes.select}
+            value={cellRadiusMode}
+            onChange={handleCellRadiusModeChange}
+            inputProps={{
+              id: 'cell-radius-mode-select',
+            }}
+          >
+            <option value="auto">Auto</option>
+            <option value="manual">Manual</option>
+          </Select>
+        </TableCell>
+      </TableRow>
+      <TableRow>
         <TableCell className={classes.labelCell}>
           {observationsLabelNice} Radius
         </TableCell>
         <TableCell className={classes.inputCell}>
           <Slider
+            disabled={cellRadiusMode !== 'manual'}
             classes={{ root: classes.slider, valueLabel: classes.sliderValueLabel }}
             value={cellRadius}
             onChange={handleRadiusChange}
             aria-labelledby="cell-radius-slider"
             valueLabelDisplay="auto"
-            step={0.25}
-            min={0.25}
-            max={8}
+            step={0.01}
+            min={0.01}
+            max={10}
           />
         </TableCell>
       </TableRow>
-      <CellColorEncodingOption
-        observationsLabel={observationsLabel}
-        cellColorEncoding={cellColorEncoding}
-        setCellColorEncoding={setCellColorEncoding}
-      />
+      <TableRow>
+        <TableCell className={classes.labelCell} htmlFor="cell-opacity-mode-select">
+          {observationsLabelNice} Opacity Mode
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Select
+            native
+            className={classes.select}
+            value={cellOpacityMode}
+            onChange={handleCellOpacityModeChange}
+            inputProps={{
+              id: 'cell-opacity-mode-select',
+            }}
+          >
+            <option value="auto">Auto</option>
+            <option value="manual">Manual</option>
+          </Select>
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          {observationsLabelNice} Opacity
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Slider
+            disabled={cellOpacityMode !== 'manual'}
+            classes={{ root: classes.slider, valueLabel: classes.sliderValueLabel }}
+            value={cellOpacity}
+            onChange={handleOpacityChange}
+            aria-labelledby="cell-opacity-slider"
+            valueLabelDisplay="auto"
+            step={0.05}
+            min={0.0}
+            max={1.0}
+          />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell} htmlFor="gene-expression-colormap-select">
+          Gene Expression Colormap
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Select
+            native
+            className={classes.select}
+            value={geneExpressionColormap}
+            onChange={handleGeneExpressionColormapChange}
+            inputProps={{
+              id: 'gene-expression-colormap-select',
+            }}
+          >
+            {GLSL_COLORMAPS.map(cmap => (
+              <option key={cmap} value={cmap}>{cmap}</option>
+            ))}
+          </Select>
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          Gene Expression Colormap Range
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Slider
+            classes={{ root: classes.slider, valueLabel: classes.sliderValueLabel }}
+            value={geneExpressionColormapRange}
+            onChange={handleColormapRangeChangeDebounced}
+            aria-labelledby="gene-expression-colormap-range-slider"
+            valueLabelDisplay="auto"
+            step={0.005}
+            min={0.0}
+            max={1.0}
+          />
+        </TableCell>
+      </TableRow>
     </OptionsContainer>
   );
 }