+++ src/layers/BitmaskLayer.js
@@ -3,6 +3,7 @@ import { project32, picking } from '@deck.gl/core'; // eslint-disable-line impor
 import { Texture2D, isWebGL2 } from '@luma.gl/core';
 import { XRLayer } from '@hms-dbmi/viv';
 import { fs, vs } from './bitmask-layer-shaders';
+import { GLSL_COLORMAPS, GLSL_COLORMAP_DEFAULT, COLORMAP_SHADER_PLACEHOLDER } from './constants';
 
 function padWithDefault(arr, defaultValue, padWidth) {
   const newArr = [...arr];
@@ -14,7 +15,9 @@ function padWithDefault(arr, defaultValue, padWidth) {
 
 const defaultProps = {
   hoveredCell: { type: 'number', value: null, compare: true },
-  cellColor: { type: 'object', value: null, compare: true },
+  cellColorData: { type: 'object', value: null, compare: true },
+  colormap: { type: 'string', value: GLSL_COLORMAP_DEFAULT, compare: true },
+  expressionData: { type: 'object', value: null, compare: true },
 };
 
 /**
@@ -24,22 +27,47 @@ const defaultProps = {
 export default class BitmaskLayer extends XRLayer {
   // eslint-disable-next-line class-methods-use-this
   getShaders() {
+    const { colormap } = this.props;
     return {
       fs,
       vs,
       modules: [project32, picking],
+      defines: {
+        [COLORMAP_SHADER_PLACEHOLDER]: GLSL_COLORMAPS.includes(colormap)
+          ? colormap
+          : GLSL_COLORMAP_DEFAULT,
+      },
     };
   }
 
   updateState({ props, oldProps, changeFlags }) {
     super.updateState({ props, oldProps, changeFlags });
-    if (props.cellColor?.data !== oldProps.cellColor?.data) {
+    if (props.cellColorData !== oldProps.cellColorData) {
       this.setColorTexture();
+    }
+    if (props.expressionData !== oldProps.expressionData) {
+      const { expressionData, cellTexHeight, cellTexWidth } = this.props;
+      const expressionTex = this.dataToTexture(
+        expressionData,
+        cellTexWidth,
+        cellTexHeight,
+      );
+      this.setState({ expressionTex });
+    }
+    if (props.colormap !== oldProps.colormap) {
+      const { gl } = this.context;
+      if (this.state.model) {
+        this.state.model.delete();
+      }
+      // eslint-disable-next-line no-underscore-dangle
+      this.setState({ model: this._getModel(gl) });
+
+      this.getAttributeManager().invalidateAll();
     }
   }
 
   setColorTexture() {
-    const { height, width, data } = this.props.cellColor;
+    const { cellColorData: data, cellTexHeight: height, cellTexWidth: width } = this.props;
     const colorTex = new Texture2D(this.context.gl, {
       width,
       height,
@@ -64,8 +92,16 @@ export default class BitmaskLayer extends XRLayer {
 
   draw(opts) {
     const { uniforms } = opts;
-    const { channelIsOn, hoveredCell } = this.props;
-    const { textures, model, colorTex } = this.state;
+    const {
+      channelIsOn,
+      hoveredCell,
+      colorScaleLo,
+      colorScaleHi,
+      isExpressionMode,
+    } = this.props;
+    const {
+      textures, model, colorTex, expressionTex,
+    } = this.state;
     // Render the image
     if (textures && model && colorTex) {
       model
@@ -73,6 +109,7 @@ export default class BitmaskLayer extends XRLayer {
           Object.assign({}, uniforms, {
             hovered: hoveredCell || 0,
             colorTex,
+            expressionTex,
             colorTexHeight: colorTex.height,
             colorTexWidth: colorTex.width,
             channelIsOn: padWithDefault(
@@ -81,6 +118,8 @@ export default class BitmaskLayer extends XRLayer {
               // There are six texture entries on the shaders
               6 - channelIsOn.length,
             ),
+            uColorScaleRange: [colorScaleLo, colorScaleHi],
+            uIsExpressionMode: isExpressionMode,
             ...textures,
           }),
         )